Fix calculateTotal function to directly add decodedPrice as a double

+++ 0x03-asynchronous/4-get_sum.dart
@@ -30,7 +30,7 @@ Future<double> calculateTotal() async {
         return -1;
       }
 
-      total += decodedPrice.toDouble();
+      total += decodedPrice;
     }
 
     return total;
